+++ Changelog.md
@@ -3,6 +3,7 @@
 # v0.0.7
 * Update Makefile.
 * Fix README.
+* Add formula parser example.
 
 # v0.0.6
 * Add time string parser.